chore(meta-service): re-organize tests for raft-store
