+++ src/meta/service/src/store/store_inner.rs
@@ -24,13 +24,13 @@ use databend_common_base::base::tokio::sync::RwLock;
 use databend_common_meta_raft_store::config::RaftConfig;
 use databend_common_meta_raft_store::key_spaces::RaftStoreEntry;
 use databend_common_meta_raft_store::leveled_store::db_exporter::DBExporter;
-use databend_common_meta_raft_store::leveled_store::leveled_map::compactor_acquirer::CompactorAcquirer;
 use databend_common_meta_raft_store::ondisk::Header;
 use databend_common_meta_raft_store::ondisk::TREE_HEADER;
 use databend_common_meta_raft_store::raft_log_v004;
 use databend_common_meta_raft_store::raft_log_v004::util;
 use databend_common_meta_raft_store::raft_log_v004::Cw;
 use databend_common_meta_raft_store::raft_log_v004::RaftLogV004;
+use databend_common_meta_raft_store::sm_v003::compactor_acquirer::CompactorAcquirer;
 use databend_common_meta_raft_store::sm_v003::write_entry::WriteEntry;
 use databend_common_meta_raft_store::sm_v003::SnapshotStoreV004;
 use databend_common_meta_raft_store::sm_v003::SMV003;
